backend/tests/test_bria_key.py
#!/usr/bin/env python3
"""
Quick test script to verify BRIA API key is configured correctly
"""

import os
import sys
from pathlib import Path
from dotenv import load_dotenv

# Load environment variables
backend_dir = Path(__file__).parent
env_path = backend_dir / ".env"
load_dotenv(env_path)

def test_bria_key():
    """Test if BRIA API key is configured and valid"""
    
    api_token = os.getenv("BRIA_API_TOKEN")
    
    if not api_token:
        print("❌ BRIA_API_TOKEN not found in .env file")
        print("\nTo set it up:")
        print("1. Edit backend/.env file")
        print("2. Add: BRIA_API_TOKEN=your_key_here")
        print("3. Or run: ./setup_api_key.sh")
        return False
    
    if api_token == "your_bria_api_token_here" or "your_" in api_token:
        print("❌ BRIA_API_TOKEN is still set to placeholder value")
        print("\nPlease update backend/.env with your actual API key")
        return False
    
    print(f"✅ BRIA_API_TOKEN found: {api_token[:10]}...{api_token[-4:]}")
    
    # Try to import and initialize client
    try:
        sys.path.insert(0, str(backend_dir))
        from core.bria_client import BRIAAPIClient
        
        client = BRIAAPIClient(api_token=api_token)
        print("✅ BRIA API client initialized successfully")
        print("\n🎉 Your API key is configured correctly!")
        print("\nYou can now:")
        print("  - Generate storyboards")
        print("  - Use image editing features")
        print("  - Generate AI animatics")
        return True
        
    except Exception as e:
        print(f"⚠️  Client initialization issue: {e}")
        print("\nThis might be normal if the API key format is correct.")
        print("Try generating a storyboard to fully test the connection.")
        return True  # Key format looks OK

if __name__ == "__main__":
    test_bria_key()




