backend/requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
pillow>=10.0.0
numpy>=1.24.0

# LLM Translation
openai>=1.0.0
anthropic>=0.3.0

# Script Parsing
python-docx>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Export Formats
reportlab>=4.0.0  # PDF
opencv-python>=4.8.0  # Video
imageio>=2.31.0  # Image I/O
imageio-ffmpeg>=0.4.9  # Video encoding
OpenEXR>=1.3.9  # HDR export
Imath>=3.1.9  # HDR support

# Utilities
pydantic>=2.0.0  # Data validation
python-dotenv>=1.0.0  # Environment variables
tqdm>=4.66.0  # Progress bars
requests>=2.31.0  # HTTP client for BRIA API

# Optional: For advanced features
scipy>=1.11.0  # Interpolation
matplotlib>=3.8.0  # Visualization

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # For FastAPI TestClient

# BRIA Local Model (Hugging Face)
huggingface-hub>=0.19.0
accelerate>=0.25.0

